+++ include/u-boot/hash-checksum.h
